requirements.txt
python-telegram-bot
openai
pytube
pydub
ffmpeg
python-dotenv
instaloader
runware
transformers
fastapi
uvicorn